\name{SCV}
\alias{SCV}
\title{Systematic Component of Variation}
\usage{
# Generalized James-Stein estimator
SCV(yi)
}
\arguments{
    \item{yi}{\code{[numeric vector]} direct estimator.}
}
\description{
Estimator of Systematic Component of Variation (SCV)
}
\references{
    McPherson, K., Wennberg, J. E., Hovind, O. B., Clifford, P. (1982).
    Small-Area Variations in the Use of Common Surgical Procedures: An
    International Comparison of New England, England, and Norway
    \emph{New England Journal of Medicine} \bold{307}, pp. 1310--1314.
    \doi{10.1056/NEJM198211183072104}
}
